requirements.txt
PyJWT>=2.0.0,<3.0.0
requests>=2.28.0,<3.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
twine>=4.0.0
build>=0.10.0
httpx>=0.24.0 